chore(ci): tentative fix pronto (#1857)
Yesterday, all our `pronto` actions started [failing like
this](https://github.com/getlago/lago-api/actions/runs/8600901566/job/23566843917?pr=1854).

It turned out that there was a bug introduced in rubocop, which made it
incompatible with `rubocop-rails`. A
[fix](https://github.com/rubocop/rubocop/pull/12826) seems on the way.

https://github.com/rubocop/rubocop/issues/12823

One way to mitigate this is to use `bundle exec rubocop` instead of
global `rubocop`. We don't have pronto in our Gemfile so let's move to
regular rubocop for now.

Ideally, we'd like to move to checks like this:
https://github.com/getlago/lago-api/pull/1856

Note: It's necessary to exit if there is no files to run because you can
run rubocop on the entire codebase today.