fix(invoice): Remove charges if charge boundaries are invalid (#1883)
https://linear.app/getlago/issue/DLVRY-262

## Context

Subscriptions fees are now invoiced at the end of the trial instead of
at the beginning of the subscription. We made a change were we don't
always skip charge fees if it's the first invoice anymore. See
https://github.com/getlago/lago-api/pull/1836

## Description

In some cases, the system might create invalid charge boundaries. This
is not new but those cases always happened on the first invoice, and
previously, we were skipping charges in this case so it was never
visible.

With this PR, we'll ensure a charge is only created if the boundaries
are valid. Next, we'll need to look into *why* the boundaries are
invalid in the first place but it's important to merge this fix in the
meantime.

### Example of invalid boundaries on upgrade

![CleanShot 2024-04-17 at 10 41
01@2x](https://github.com/getlago/lago-api/assets/1525636/2ae49ab7-7fff-497c-8290-608c91b47e87)