feat (accounting-integrations): add logic for fetching items from integration aggregator (#1890)
## Context

Currently Lago does not support accounting integrations

## Description

This PR adds logic for fetching items from Nango (Lago's integration
aggregator).

These items are basically objects that are used for mapping. We decided
to store it on our side in order to make it searchable and more
efficient since the same list need to be returned to the UI for various
mappable objects.