chore(ci): Migrate DB before test instead of running a dedicated job (#1888)
## Description

Recently, a [new Action was set
up](https://github.com/getlago/lago-api/pull/1761) to test DB
migrations. I agree 100% that the migrations should be tested as part of
the CI.

I propose to run the migration when setup the Actions to run rspec
instead of loading the schema.

A small benefit is that we don't need to setup a dedicated env "just" to
run the migrations.

A big benefit is that you can only commit `db/schema.rb` when your PR is
ready and still run the CI in the meantime. `schema.rb` will often cause
conflicts in PR that stays open for while. We aim for short-lived PR of
course, but sometimes some stay open longer.