feat (accounting-integrations): add logic and graphql support for fetching subsidiaries list (#1935)
## Context

Currently Lago does not support accounting integrations

## Description

This PR covers fetching subsidiaries from Lago's integration aggregator.

List of subsidiaries is needed on the UI in order to allow user to
select customer's subsidiary. Subsidiary is required field in customer
create and update actions.

There is small amount of subsidiaries (and it is used only on one place
in the UI) so we will return it to the UI on the fly from accounting
aggregator.