feat(accounting-integrations): add strategy service for creating or updating integration customers (#1990)
## Context

Currently Lago does not support accounting integrations

## Description

This PR handles strategy service for accounting customers.

This service will be called from `customers/create_service.rb` and
`customers/update_service.rb`.

The main goal of this service is to determine if we should:
- skip syncing accounting customer
- create integration customer
- update integration customer (remove one)

Based on decision we call appropriate job so that the whole logic of
accounting customer is extracted and can be retried. Accounting customer
does not need to be in the same DB transaction and the main goal was
that the whole process which handles accounting customer does not add
extra complexity in customer create/update services.